+++ packages/wdio-cli/src/utils.ts
@@ -358,6 +358,11 @@ export function hasBabelConfig(rootDir: string) {
  * detect if project has a compiler file
  */
 export async function detectCompiler(answers: Questionnair) {
+    // obviously there is no compiler, if there is no package.json
+    if (answers.createPackageJSON) {
+        return false
+    }
+
     const root = await getProjectRoot(answers)
     const rootTSConfigExist = await fs.access(path.resolve(root, 'tsconfig.json')).then(() => true, () => false)
     return (await hasBabelConfig(root) || rootTSConfigExist) ? true : false
@@ -533,42 +538,41 @@ export async function getAnswers(yes: boolean): Promise<Questionnair> {
 
     const projectProps = await getProjectProps(process.cwd())
     const isProjectExisting = Boolean(projectProps)
+    const nameInPackageJsonIsNotCreateWdioDefault = projectProps?.packageJson?.name !== 'my-new-project'
     const projectName = projectProps?.packageJson?.name ? ` named "${projectProps.packageJson.name}"` : ''
-    const questions = [
-        /**
-         * in case the `wdio config` was called using a global installed @wdio/cli package
-         */
-        ...(!isProjectExisting
-            ? [{
-                type: 'confirm',
-                name: 'createPackageJSON',
-                default: true,
-                message: `Couldn't find a package.json in "${process.cwd()}" or any of the parent directories, do you want to create one?`,
-            }]
-            /**
-             * in case create-wdio was used which creates a package.json with name "my-new-project"
-             * we don't need to ask this question
-             */
-            : projectProps?.packageJson?.name !== 'my-new-project'
-                ? [{
-                    type: 'confirm',
-                    name: 'projectRootCorrect',
-                    default: true,
-                    message: `A project${projectName} was detected at "${projectProps?.path}", correct?`,
-                }, {
-                    type: 'input',
-                    name: 'projectRoot',
-                    message: 'What is the project root for your test project?',
-                    default: projectProps?.path,
-                    // only ask if there are more than 1 runner to pick from
-                    when: /* istanbul ignore next */ (answers: Questionnair) => !answers.projectRootCorrect
-                }]
-                : []
-        ),
-        ...QUESTIONNAIRE
+
+    const projectRootQuestions = [
+        {
+            type: 'confirm',
+            name: 'createPackageJSON',
+            default: true,
+            message: `Couldn't find a package.json in "${process.cwd()}" or any of the parent directories, do you want to create one?`,
+            when: !isProjectExisting
+        },
+        {
+            type: 'confirm',
+            name: 'projectRootCorrect',
+            default: true,
+            message: `A project${projectName} was detected at "${projectProps?.path}", correct?`,
+            when: isProjectExisting && nameInPackageJsonIsNotCreateWdioDefault
+        }, {
+            type: 'input',
+            name: 'projectRoot',
+            message: 'What is the project root for your test project?',
+            default: projectProps?.path,
+            // only ask if there are more than 1 runner to pick from
+            when: /* istanbul ignore next */ (answers: Questionnair) => isProjectExisting && nameInPackageJsonIsNotCreateWdioDefault && !answers.projectRootCorrect
+        }
     ]
+
     // @ts-expect-error
-    return inquirer.prompt(questions)
+    const answers = await inquirer.prompt(projectRootQuestions)
+    if (answers.createPackageJSON) {
+        answers.projectRoot = process.cwd()
+    }
+
+    // @ts-expect-error
+    return inquirer.prompt(QUESTIONNAIRE, answers)
 }
 
 /**
@@ -995,6 +999,10 @@ export async function createWDIOConfig(parsedAnswers: ParsedAnswers) {
  * @returns project root path
  */
 export async function getProjectRoot (parsedAnswers?: Questionnair) {
+    if (parsedAnswers?.createPackageJSON && parsedAnswers.projectRoot) {
+        return parsedAnswers.projectRoot
+    }
+
     const root = (await getProjectProps())?.path
     if (!root) {
         throw new Error('Could not find project root directory with a package.json')