+++ packages/webdriverio/src/commands/browser/executeAsync.ts
@@ -4,8 +4,8 @@ import type { remote } from 'webdriver'
 import { verifyArgsAndStripIfElement } from '../../utils/index.js'
 import { LocalValue } from '../../utils/bidi/value.js'
 import { parseScriptResult } from '../../utils/bidi/index.js'
-import { getContextManager } from '../../context.js'
-import { NAME_POLYFILL } from '../../polyfill.js'
+import { getContextManager } from '../../session/context.js'
+import { polyfillFn } from '../../session/polyfill.js'
 
 /**
  * :::warning
@@ -103,7 +103,8 @@ export async function executeAsync<ReturnValue, InnerArguments extends unknown[]
      */
     if (typeof script === 'function') {
         script = `
-            ${NAME_POLYFILL}
+            ${polyfillFn}
+            webdriverioPolyfill()
             return (${script}).apply(null, arguments)
         `
     }