+++ packages/webdriverio/tests/shadowRoot.test.ts
@@ -1,11 +1,12 @@
 import { describe, it, vi, expect, beforeEach } from 'vitest'
 
-import { getShadowRootManager, ShadowRootTree } from '../src/shadowRoot.js'
+import { getShadowRootManager, ShadowRootTree } from '../src/session/shadowRoot.js'
 
 const defaultBrowser = {
     sessionSubscribe: vi.fn().mockResolvedValue({}),
     on: vi.fn(),
     scriptAddPreloadScript: vi.fn(),
+    capabilities: {}
 }
 
 describe('ShadowRootManager', () => {
@@ -26,12 +27,12 @@ describe('ShadowRootManager', () => {
     it('registers correct event listeners', async () => {
         const wid = process.env.WDIO_UNIT_TESTS
         delete process.env.WDIO_UNIT_TESTS
-        const browser = { ...defaultBrowser, isBidi: true, options: { automationProtocol: 'webdriver' } } as any
+        const browser = { ...defaultBrowser, isBidi: true, options: { capabilities: { webSocketUrl: './' } } } as any
         const manager = getShadowRootManager(browser)
         process.env.WDIO_UNIT_TESTS = wid
         expect(await manager.initialize()).toBe(true)
         expect(browser.sessionSubscribe).toBeCalledTimes(1)
-        expect(browser.on).toBeCalledTimes(3)
+        expect(browser.on).toBeCalledTimes(4)
         expect(browser.scriptAddPreloadScript).toBeCalledTimes(1)
     })
 
@@ -40,7 +41,7 @@ describe('ShadowRootManager', () => {
         const manager = getShadowRootManager(browser)
         expect(await manager.initialize()).toBe(true)
         expect(browser.sessionSubscribe).toBeCalledTimes(0)
-        expect(browser.on).toBeCalledTimes(0)
+        expect(browser.on).toBeCalledTimes(1)
         expect(browser.scriptAddPreloadScript).toBeCalledTimes(0)
     })
 
@@ -49,7 +50,7 @@ describe('ShadowRootManager', () => {
         const manager = getShadowRootManager(browser)
         expect(await manager.initialize()).toBe(true)
         expect(browser.sessionSubscribe).toBeCalledTimes(0)
-        expect(browser.on).toBeCalledTimes(0)
+        expect(browser.on).toBeCalledTimes(1)
         expect(browser.scriptAddPreloadScript).toBeCalledTimes(0)
     })